refactor: move all dropdowns to same directory
