+++ src/components/Configs/ConfigsListFilters/ConfigTypesTriStateDropdown.tsx
@@ -4,7 +4,7 @@ import {
 } from "@flanksource-ui/api/services/configs";
 import TristateReactSelect, {
   TriStateOptions
-} from "@flanksource-ui/ui/TristateReactSelect/TristateReactSelect";
+} from "@flanksource-ui/ui/Dropdowns/TristateReactSelect";
 import { useQuery } from "@tanstack/react-query";
 import { useCallback } from "react";
 import { useSearchParams } from "react-router-dom";
