+++ src/components/JobsHistory/Filters/JobHistoryNames.tsx
@@ -1,6 +1,6 @@
 import TristateReactSelect, {
   TriStateOptions
-} from "@flanksource-ui/ui/TristateReactSelect/TristateReactSelect";
+} from "@flanksource-ui/ui/Dropdowns/TristateReactSelect";
 import { formatJobName } from "@flanksource-ui/utils/common";
 import { useQuery } from "@tanstack/react-query";
 import { useSearchParams } from "react-router-dom";
