+++ src/components/JobsHistory/Filters/JobsHistoryFilters.tsx
@@ -1,8 +1,8 @@
-import { TimeRangePicker } from "@flanksource-ui/ui/TimeRangePicker";
-import useTimeRangeParams from "@flanksource-ui/ui/TimeRangePicker/useTimeRangeParams";
 import TristateReactSelect, {
   TriStateOptions
-} from "@flanksource-ui/ui/TristateReactSelect/TristateReactSelect";
+} from "@flanksource-ui/ui/Dropdowns/TristateReactSelect";
+import { TimeRangePicker } from "@flanksource-ui/ui/TimeRangePicker";
+import useTimeRangeParams from "@flanksource-ui/ui/TimeRangePicker/useTimeRangeParams";
 import dayjs from "dayjs";
 import { URLSearchParamsInit, useSearchParams } from "react-router-dom";
 import { ReactSelectDropdown, StateOption } from "../../ReactSelectDropdown";
